extract getSuspendedSale
